refactor(mkrootfs): use sha512crypt instead of crypt(old)
I misunderstood the meaning of Perl's crypt function,
I thought the second argument was a hash method, but
it was actually salt. using crypt gives the result of
a traditional Unix password hash function, which is
very insecure. Arch Linux currently uses sha512crypt
by default, and we use it as well.

Signed-off-by: Coelacanthus <CoelacanthusHex@gmail.com>